review fix from #8600
