+++ ocs_ci/ocs/cluster.py
@@ -3060,22 +3060,6 @@ def get_mds_standby_replay_info():
     }
 
 
-def set_osd_op_complaint_time(osd_op_complaint_time_val: float) -> dict:
-    """
-    Set osd_op_complaint_time to the given value
-
-    Args:
-        osd_op_complaint_time_val (float): Value in seconds to set osd_op_complaint_time to
-
-    Returns:
-        dict: output of the command
-    """
-    ct_pod = pod.get_ceph_tools_pod()
-    return ct_pod.exec_ceph_cmd(
-        f"ceph config set osd osd_op_complaint_time {osd_op_complaint_time_val}"
-    )
-
-
 def get_full_ratio_from_osd_dump():
     """
     Get the full ratio value from osd map
